+++ R/as_mapper.R
@@ -113,3 +113,30 @@ plucker <- function(i, default) {
     env = caller_env()
   )
 }
+
+as_predicate <- function(.fn, ..., .mapper) {
+  if (.mapper) {
+    .fn <- as_mapper(.fn, ...)
+  }
+
+  function(...) {
+    out <- .fn(...)
+
+    if (!is_bool(out)) {
+      abort(sprintf(
+        "Predicate functions must return a single logical `TRUE` or `FALSE`, not %s",
+        as_predicate_friendly_type_of(out)
+      ))
+    }
+
+    out
+  }
+}
+
+as_predicate_friendly_type_of <- function(x) {
+  if (is_na(x)) {
+    "a missing value"
+  } else {
+    friendly_type_of(x, length = TRUE)
+  }
+}