+++ R/every-some.R
@@ -17,23 +17,27 @@
 #' y %>% every(is.numeric)
 #' y %>% every(is.integer)
 every <- function(.x, .p, ...) {
-  .p <- as_mapper(.p, ...)
+  .p <- as_predicate(.p, ..., .mapper = TRUE)
+
   for (i in seq_along(.x)) {
     val <- .p(.x[[i]], ...)
     if (is_false(val)) return(FALSE)
     if (anyNA(val)) return(NA)
   }
+
   TRUE
 }
 
 #' @export
 #' @rdname every
 some <- function(.x, .p, ...) {
-  .p <- as_mapper(.p, ...)
+  .p <- as_predicate(.p, ..., .mapper = TRUE)
+
   val <- FALSE
   for (i in seq_along(.x)) {
     val <- val || .p(.x[[i]], ...)
     if (is_true(val)) return(TRUE)
   }
+
   val
 }