+++ tests/testthat/test-detect.R
@@ -22,3 +22,8 @@ test_that("has_element checks whether a list contains an object", {
   expect_true(has_element(list(1, 2), 1))
   expect_false(has_element(list(1, 2), 3))
 })
+
+test_that("`detect()` requires a predicate function", {
+  expect_error(detect(list(1:2, 2), is.na), "must return a single logical")
+  expect_error(detect(list(1:2, 2), function(...) NA), "not a missing value")
+})